Metadata mapping on type in stead of name
