+++ CPS-API/Models/FileMetadata.cs
@@ -1,4 +1,5 @@
-﻿using System.Text.Json.Serialization;
+﻿using System.Globalization;
+using System.Text.Json.Serialization;
 
 namespace CPS_API.Models
 {
@@ -38,10 +39,41 @@ namespace CPS_API.Models
             set
             {
                 var property = this.GetType().GetProperty(fieldname);
-                if (property != null)
-                    property.SetValue(this, value, null);
+                if (property == null) throw new ArgumentException("Unknown property " + fieldname);
+
+                if (property.PropertyType == typeof(Classification))
+                {
+                    var stringValue = value?.ToString();
+                    if (stringValue != null)
+                    {
+                        Enum.TryParse<Classification>(stringValue, out var enumValue);
+                        property.SetValue(this, enumValue, null);
+                    }
+                }
+                else if (property.PropertyType == typeof(int))
+                {
+                    var stringValue = value?.ToString();
+                    var decimalValue = Convert.ToDecimal(stringValue, new CultureInfo("en-US"));
+                    if (decimalValue % 1 == 0)
+                    {
+                        property.SetValue(this, (int)decimalValue, null);
+                    }
+                }
+                else if (property.PropertyType == typeof(DateTime))
+                {
+                    var stringValue = value?.ToString();
+                    DateTime.TryParse(stringValue, out var dateValue);
+                    property.SetValue(this, dateValue, null);
+                }
+                else if (property.PropertyType == typeof(string))
+                {
+                    var stringValue = value?.ToString();
+                    property.SetValue(this, stringValue, null);
+                }
                 else
-                    throw new ArgumentException("Unknown property " + fieldname);
+                {
+                    property.SetValue(this, value, null);
+                }
             }
         }
     }
